requirements.txt
azure-ai-projects
azure-identity
azure-ai-evaluation
pyyaml
pytest
flake8